cookies and sessions ready

+++ views/index.tpl
@@ -17,8 +17,11 @@
                     <?=$film['title']?>
                 </h4>
                 <div class="buttons">
-                    <a href="edit.php?id=<?=$film['id']?>" class="button button--edit">Редактировать</a>
-                    <a href="?action=delete&id=<?=$film['id']?>" class="button button--remove">Удалить</a>
+                <?php if( isset($_SESSION['user'])) {
+                    if($_SESSION['user'] == 'admin') { ?>
+                        <a href="edit.php?id=<?=$film['id']?>" class="button button--edit">Редактировать</a>
+                        <a href="?action=delete&id=<?=$film['id']?>" class="button button--remove">Удалить</a>
+                <?php } } ?>                
                 </div>
             </div>
             <div class="badge">
